requirements.txt
msgspec~=0.18.6
pyzmq~=26.1.0
pytest~=8.3.2